@@ -1,10 +1,9 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace TheNoobs.DependencyInjection.Extensions.Modules.Abstractions
+namespace TheNoobs.DependencyInjection.Extensions.Modules.Abstractions;
+
+public interface IServiceModuleSetup
 {
-    public interface IServiceModuleSetup
-    {
-        void Setup(IServiceCollection services, IConfiguration configuration);  
-    }
+    void Setup(IServiceCollection services, IConfiguration configuration);  
 }
