@@ -8,85 +8,84 @@ using Moq;
 using TheNoobs.DependencyInjection.Extensions.Modules.Abstractions;
 using Xunit;
 
-namespace TheNoobs.DependencyInjection.Extensions.Modules.UnitTests
+namespace TheNoobs.DependencyInjection.Extensions.Modules.UnitTests;
+
+public class DependencyInjectionExtensionsTests
 {
-    public class DependencyInjectionExtensionsTests
+    [Fact]
+    public void GivenDependencyInjectionExtensionsWhenAddModuleShouldSetup()
     {
-        [Fact]
-        public void GivenDependencyInjectionExtensionsWhenAddModuleShouldSetup()
-        {
-            IConfiguration configuration = new ConfigurationBuilder().Build();
-            IServiceCollection services = new ServiceCollection();
-            services.AddInjections(configuration, typeof(DependencyInjectionExtensionsTests).Assembly);
+        IConfiguration configuration = new ConfigurationBuilder().Build();
+        IServiceCollection services = new ServiceCollection();
+        services.AddInjections(configuration, typeof(DependencyInjectionExtensionsTests).Assembly);
 
-            using var provider = services.BuildServiceProvider();
-            var testClass = provider.GetService<ModuleTestClass>();
+        using var provider = services.BuildServiceProvider();
+        var testClass = provider.GetService<ModuleTestClass>();
 
-            testClass.Should().NotBeNull();
-            testClass.Should().BeOfType<ModuleTestClass>();
-        }
+        testClass.Should().NotBeNull();
+        testClass.Should().BeOfType<ModuleTestClass>();
+    }
         
-        [Fact]
-        public void GivenDependencyInjectionExtensionsWhenUseModuleShouldSetup()
-        {
-            var servicesMock = new Mock<IServiceProvider>();
-            var appBuilder = new Mock<IApplicationBuilder>();
+    [Fact]
+    public void GivenDependencyInjectionExtensionsWhenUseModuleShouldSetup()
+    {
+        var servicesMock = new Mock<IServiceProvider>();
+        var appBuilder = new Mock<IApplicationBuilder>();
             
-            appBuilder.Setup(x => x.ApplicationServices).Returns(servicesMock.Object)
-                .Verifiable();
-            servicesMock.Setup(x => x.GetService(typeof(ModuleTestClass)))
-                .Returns(new ModuleTestClass())
-                .Verifiable();
+        appBuilder.Setup(x => x.ApplicationServices).Returns(servicesMock.Object)
+            .Verifiable();
+        servicesMock.Setup(x => x.GetService(typeof(ModuleTestClass)))
+            .Returns(new ModuleTestClass())
+            .Verifiable();
 
-            appBuilder.Object.UseInjections(typeof(DependencyInjectionExtensionsTests).Assembly);
+        appBuilder.Object.UseInjections(typeof(DependencyInjectionExtensionsTests).Assembly);
 
-            appBuilder.VerifyAll();
-            servicesMock.VerifyAll();
-        }
+        appBuilder.VerifyAll();
+        servicesMock.VerifyAll();
+    }
 
-        private class ModuleTestClass
+    private class ModuleTestClass
+    {
+    }
+
+    private class ServiceModuleSetup : IServiceModuleSetup
+    {
+        public void Setup(IServiceCollection services, IConfiguration configuration)
         {
+            services.AddScoped<ModuleTestClass>();
         }
-
-        private class ServiceModuleSetup : IServiceModuleSetup
-        {
-            public void Setup(IServiceCollection services, IConfiguration configuration)
-            {
-                services.AddScoped<ModuleTestClass>();
-            }
-        }
+    }
         
-        private class AppBuilderApplicationModuleSetup : IApplicationModuleSetup
+    private class AppBuilderApplicationModuleSetup : IApplicationModuleSetup
+    {
+        public void Setup(IApplicationBuilder applicationBuilder)
         {
-            public void Setup(IApplicationBuilder applicationBuilder)
-            {
-                var _ = applicationBuilder.ApplicationServices.GetRequiredService<ModuleTestClass>();
-            }
+            var _ = applicationBuilder.ApplicationServices.GetRequiredService<ModuleTestClass>();
+        }
+    }
+
+    private class OrderClassTest
+    {
+    }
+
+    private class Module1Setup : IServiceModuleSetup, IOrderedModule
+    {
+        public void Setup(IServiceCollection services, IConfiguration configuration)
+        {
+            services.Any(x => x.ServiceType == typeof(OrderClassTest)).Should().BeTrue();
         }
 
-        private class OrderClassTest
-        {
-        }
-
-        private class Module1Setup : IServiceModuleSetup, IOrderedModule
-        {
-            public void Setup(IServiceCollection services, IConfiguration configuration)
-            {
-                services.Any(x => x.ServiceType == typeof(OrderClassTest)).Should().BeTrue();
-            }
-
-            public int Order => 2;
-        }
+        public int Order => 2;
+    }
         
-        private class Module2Setup : IServiceModuleSetup, IOrderedModule
+    private class Module2Setup : IServiceModuleSetup, IOrderedModule
+    {
+        public void Setup(IServiceCollection services, IConfiguration configuration)
         {
-            public void Setup(IServiceCollection services, IConfiguration configuration)
-            {
-                services.Any(x => x.ServiceType == typeof(OrderClassTest)).Should().BeFalse();
-                services.AddScoped<OrderClassTest>();
-            }
-
-            public int Order => 1;
+            services.Any(x => x.ServiceType == typeof(OrderClassTest)).Should().BeFalse();
+            services.AddScoped<OrderClassTest>();
         }
+
+        public int Order => 1;
     }
 }
